% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{all.funs}
\alias{all.funs}
\title{Get all functions that are used in a formula  \code{expr}.}
\usage{
\method{all}{funs}(expr)
}
\arguments{
\item{expr}{}
}
\value{

}
\description{
Get all functions that are used in a formula  \code{expr}.
}
\examples{

}
