% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{all_funs}
\alias{all_funs}
\title{Get all functions that are used in a formula  \code{expr}.}
\usage{
all_funs(expr)
}
\arguments{
\item{expr}{a formula object}
}
\value{

}
\description{
Get all functions that are used in a formula  \code{expr}.
}
\examples{

}
