% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn2dagitty.R
\name{bn2dagitty}
\alias{bn2dagitty}
\title{Converts a bn_df object to a dagitty object}
\usage{
bn2dagitty(bn_df)
}
\arguments{
\item{bn_df}{initialised bn_df object, with simulation instructions. Created with \code{bn_create}}
}
\value{
dagitty object
}
\description{
Converts a bn_df object to a dagitty object
}
\examples{

}
