% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_simulate.R
\name{bn_simulate}
\alias{bn_simulate}
\title{Simulate data from bn_df object}
\usage{
bn_simulate(bn_df, data = NULL, pop_size, keep_all = FALSE)
}
\arguments{
\item{keep_all}{}
}
\value{

}
\description{
Simulate data from bn_df object
}
