% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfactor.R
\name{rfactor}
\alias{rfactor}
\title{Random factor variables}
\usage{
rfactor(n, levels, p)
}
\arguments{
\item{n}{number of samples}

\item{levels}{vector of categories to sample from}

\item{p}{vector of probabilities}
}
\value{
a \code{factor} vector
}
\description{
Random factor variables
}
\examples{
#' rfactor(n=10, levels=c("a","b"), p=c(0.2,0.8))
}
