% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_create.R
\name{bn_create}
\alias{bn_create}
\title{Creates a bayesian network object from a list of nodes}
\usage{
bn_create(list, known_variables = NULL)
}
\arguments{
\item{list}{of node objects, created by \code{node}.}

\item{known_variables}{}
}
\value{
data.frame
}
\description{
Converts list to data frame which is a bit easier to work with, and embellishes with some useful columns.
The function performs a few checks on the the list, for instance to make sure the p-DAG is indeed acyclic
and that variables using in the expressions are defined elsewhere.
The known_variables argument is for passing a character vector of variables names
for variables that are already defined externally in a
given dataset, which can be passed to bn_simulate
whilst variable_formula is the variable name itself, this is to help with the bn_simulate function
it doesn't actually lead to self-dependence (eg var depends on var)
}
