% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_plot.R
\name{bn_plot}
\alias{bn_plot}
\title{Plot bn_df object}
\usage{
bn_plot(bn_df)
}
\arguments{
\item{bn_df}{initialised bn_df object, with simulation instructions. Created with \code{bn_create}}
}
\value{

}
\description{
Plot bn_df object
}
