% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_simulate.R
\name{bn_simulate}
\alias{bn_simulate}
\title{Simulate data from bn_df object}
\usage{
bn_simulate(bn_df, known_df = NULL, pop_size, keep_all = FALSE, .id = NULL)
}
\arguments{
\item{bn_df}{initialised bn_df object, with simulation instructions. Created with \code{bn_create}}

\item{known_df}{data.frame. Optional data.frame containing upstream variables used for simulation.}

\item{pop_size}{integer. The size of the dataset to be created.}

\item{keep_all}{logical. Keep all simulated variables or only keep those specified by \code{keep}}

\item{.id}{character. Name of id column placed at the start of the dataset. If NULL (default) then no id column is created.}
}
\value{
tbl
}
\description{
Simulate data from bn_df object
}
